improvement #166: split trading loop processings from trade_rules
