+++ src/swing_trader.py
@@ -2,7 +2,6 @@ from typing import Any, Dict, Union
 
 import pandas as pd
 
-import src.trade_rules.base as base_rules
 import src.trade_rules.stoploss as stoploss_strategy
 from src.trader import Trader
 
@@ -19,11 +18,12 @@ class SwingTrader(Trader):
     ) -> Dict[str, Union[str, pd.DataFrame]]:
         """backtest swing trade"""
         self.__generate_entry_column(candles=candles)
-        sliding_result = self.__slide_to_reasonable_prices(candles=candles)
+        result = self.__slide_to_reasonable_prices(candles=candles)
 
-        candles.to_csv("./tmp/csvs/full_data_dump.csv")
-        result_msg: str = self.__result_message(sliding_result["result"])
-        return {"result": result_msg, "candles": candles}
+        return {
+            "result": result["result"],
+            "candles": candles,
+        }
 
     # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
     # Private
@@ -32,9 +32,9 @@ class SwingTrader(Trader):
         print("[Trader] judging entryable or not ...")
 
         entry_direction: pd.Series = candles["entryable"].fillna(method="ffill")
-        candles_with_stoploss: pd.DataFrame = self.__set_stoploss_prices(candles, entry_direction)
-        base_rules.commit_positions(
-            candles_with_stoploss,
+        candles.loc[:, "possible_stoploss"] = self.__set_stoploss_prices(candles, entry_direction)
+        self._commit_positions(
+            candles,
             long_indexes=(entry_direction == "long"),
             short_indexes=(entry_direction == "short"),
             spread=self.config.static_spread,
@@ -43,10 +43,11 @@ class SwingTrader(Trader):
     def __set_stoploss_prices(
         self, candles: pd.DataFrame, entry_direction: pd.Series
     ) -> pd.DataFrame:
-        candles.loc[:, "possible_stoploss"] = stoploss_strategy.previous_candle_othersides(
-            candles, entry_direction, self.config
+        return stoploss_strategy.previous_candle_othersides(
+            candles,
+            entry_direction,
+            self.config,
         )
-        return candles
 
     # OPTIMIZE: probably this method has many unnecessary processings!
     def __slide_to_reasonable_prices(self, candles: pd.DataFrame) -> Dict[str, str]:
@@ -59,7 +60,6 @@ class SwingTrader(Trader):
             "records"
         )
         if position_rows == []:
-            print("[Trader] no positions ...")
             return {"result": "no position"}
 
         df_with_positions = pd.DataFrame.from_dict(position_rows)
@@ -71,11 +71,50 @@ class SwingTrader(Trader):
             df_with_positions["time"].astype(str).to_numpy(copy=True)
         )
 
-        print("[Trader] finished sliding !")
-        return {"result": "success"}
+        return {"result": "[Trader] 1 series of trading is FINISHED!"}
 
-    def __result_message(self, result: str) -> str:
-        if result == "no position":
-            return "no position"
+    def _commit_positions(
+        self,
+        candles: pd.DataFrame,
+        long_indexes: pd.Series,
+        short_indexes: pd.Series,
+        spread: float,
+    ) -> None:
+        """
+        set timing and price of exit
 
-        return "[Trader] 1 series of trading is FINISHED!"
+        Parameters
+        ----------
+        candles : pd.DataFrame
+            Index:
+                Any
+            Columns:
+                Name: high,              dtype: float64 (required)
+                Name: low,               dtype: float64 (required)
+                Name: entryable,         dtype: object  (required)
+                Name: entryable_price,   dtype: float64 (required)
+                Name: possible_stoploss, dtype: float64 (required)
+                Name: time,              dtype: object  # datetime64[ns]
+
+        Returns
+        -------
+        None
+        """
+        candles.loc[:, "position"] = candles["entryable"].copy()
+
+        long_exits = long_indexes & (candles["low"] < candles["possible_stoploss"])
+        candles.loc[long_exits, "position"] = "sell_exit"
+        candles.loc[long_exits, "exitable_price"] = candles.loc[long_exits, "possible_stoploss"]
+
+        short_exits = short_indexes & (candles["high"] + spread > candles["possible_stoploss"])
+        candles.loc[short_exits, "position"] = "buy_exit"
+        candles.loc[short_exits, "exitable_price"] = candles.loc[short_exits, "possible_stoploss"]
+
+        # INFO: position column の整理
+        candles["position"].fillna(method="ffill", inplace=True)
+
+        # INFO: 2連続entry, entryなしでのexitを除去
+        no_position_index = (candles["position"] == candles["position"].shift(1)) & (
+            candles["entryable_price"].isna() | candles["exitable_price"].isna()
+        )
+        candles.loc[no_position_index, "position"] = None
